+++ src/pages/orders/all-orders/components/OrdersTable/index.js
@@ -1,0 +1,84 @@
+import { connect } from 'dva';
+import {
+    Table,
+} from 'antd';
+import moment from 'moment';
+const mapStateToProps = ({ orders, loading }) => ({
+    ordersData: orders.ordersData,
+    loading: loading.models["orders"],
+})
+const mapDispatchToProps = (dispatch) => ({
+    getOrders: () => dispatch({
+        type: 'orders/setOrders_e'
+    }),
+})
+@connect(mapStateToProps, mapDispatchToProps)
+export default class OrdersTable extends React.Component {
+    componentDidMount() {
+        const { getOrders } = this.props;
+        getOrders();
+    }
+    render() {
+        const { ordersData, loading, } = this.props;
+        const columns = [
+            {
+                title: 'Order',
+                dataIndex: 'name',
+                key: 'name',
+            },
+            {
+                title: 'Date',
+                dataIndex: 'created_at',
+                key: 'created_at',
+                render: created_at => (moment(created_at).format("YYYY-MM-DD HH:mm:ss"))
+            },
+            {
+                title: 'Customer',
+                dataIndex: 'customer',
+                key: 'customer',
+                render: customer => (customer ? customer.first_name + " " + customer.last_name : "没有客户"),
+            },
+            {
+                title: 'Payment',
+                dataIndex: 'financial_status',
+                key: 'financial_status',
+            },
+            {
+                title: 'Fulfillment',
+                dataIndex: 'fulfillment_status',
+                key: 'fulfillment_status',
+                render: fulfillment_status => {
+                    if (fulfillment_status === null) {
+                        return 'Unfulfilled'
+                    }
+                    else {
+                        return fulfillment_status
+                    }
+                },
+            },
+            {
+                title: 'Total',
+                dataIndex: 'total_line_items_price',
+                key: 'total_line_items_price',
+                render: total_line_items_price => ('$' + total_line_items_price)
+            },
+        ];
+        const rowSelection = {
+            onChange: (selectedRowKeys, selectedRows) => {
+                // console.log(`selectedRowKeys: ${selectedRowKeys}`, 'selectedRows: ', selectedRows);
+            },
+            getCheckboxProps: record => ({
+                // disabled: record.customer.first_name === '', // Column configuration not to be checked
+            }),
+        };
+        return (
+            <Table
+                columns={columns}
+                dataSource={ordersData}
+                rowSelection={rowSelection}
+                pagination={false}
+                loading={loading}
+            />
+        );
+    }
+}
