+++ src/services/shopConfig.js
@@ -1,4 +1,3 @@
-
 export const shopName = 'maqii'; //我的商店名字
 export const apiBase = `https://mirror.viralbox.org/${shopName}/admin/api/2019-10`; //api根
 export const access_token='6d699c307bc71730981e6eaedba967cd'; //认证
